+++ examples/heatmaps/BlockHeatMapChart.tsx
@@ -14,7 +14,7 @@ import {
 import {
     HeatMap,
     HeatMapCells,
-    HeatMapColorRectangle,
+    HeatMapRectangle,
     HeatMapHighlight,
     isHeatMapData,
 } from '@chsk/matrix'
@@ -118,7 +118,7 @@ export const BlockHeatMapChart = ({ fref, chartData, rawData }: MilestoneStory) 
                 }}
             >
                 <MilestoneMotion enterOn={'map'}>
-                    <HeatMapCells cell={HeatMapColorRectangle} />
+                    <HeatMapCells cell={HeatMapRectangle} />
                 </MilestoneMotion>
                 <Surface style={{ stroke: '#222222', strokeWidth: 1, fill: '#ffffff00' }} />
                 <Axis variant={'left'} label={'Replicate 1'} />