+++ examples/heatmaps/DivergingHeatMapChart.tsx
@@ -9,7 +9,7 @@ import {
     Tooltip,
     TooltipProvider,
 } from '@chsk/core'
-import { HeatMap, HeatMapCells, HeatMapColorRectangle, HeatMapHighlight } from '@chsk/matrix'
+import { HeatMap, HeatMapCells, HeatMapRectangle, HeatMapHighlight } from '@chsk/matrix'
 import { interpolateBrBG } from 'd3-scale-chromatic'
 import { generateHeatMapMatrixNormal } from './generators'
 import { alphabetGreek } from '../utils'
@@ -40,7 +40,7 @@ export const DivergingHeatMapChart = ({ fref, chartData, rawData }: MilestoneSto
             >
                 <HeatMapCells
                     style={{ strokeWidth: 0.5, stroke: '#888888' }}
-                    cell={HeatMapColorRectangle}
+                    cell={HeatMapRectangle}
                 />
                 <Axis variant={'left'}>
                     <AxisTicks