+++ examples/heatmaps/TriangularHeatMapChart.tsx
@@ -13,7 +13,7 @@ import {
 import {
     HeatMap,
     HeatMapCells,
-    HeatMapColorRectangle,
+    HeatMapRectangle,
     HeatMapHighlight,
     isHeatMapData,
 } from '@chsk/matrix'
@@ -106,7 +106,7 @@ export const TriangularHeatMapChart = ({ fref, chartData, rawData }: MilestoneSt
                     domain: [0, 100],
                 }}
             >
-                <HeatMapCells cells={lowerTriangularCells} cell={HeatMapColorRectangle} />
+                <HeatMapCells cells={lowerTriangularCells} cell={HeatMapRectangle} />
                 <HeatMapHighlight style={{ fill: '#222222', opacity: 0.6 }} />
                 <HeatMapCells
                     cells={upperTriangularCells}