+++ packages/matrix/stories/heatmaps/HeatMapSimpleRectangle.stories.tsx
@@ -1,19 +1,19 @@
-import { HeatMapCells, HeatMapColorCircle } from '../../src'
+import { HeatMapCells, HeatMapRectangle } from '../../src'
 import { ChartHeatMapDecorator } from '../decorators'
 
 export default {
-    title: 'Addons/Matrix/HeatMaps/HeatMapColorCircle',
+    title: 'Addons/Matrix/HeatMaps/HeatMapSimpleRectangle',
     component: HeatMapCells,
 }
 
 export const Example = {
-    name: 'circles',
+    name: 'rectangle',
     args: {
         style: {
             strokeWidth: 1,
             stroke: '#222222',
         },
-        cell: HeatMapColorCircle,
+        cell: HeatMapRectangle,
     },
     component: HeatMapCells,
     decorators: [ChartHeatMapDecorator],